requirements.txt
fastmcp
requests
yfinance